Change ignored names to regex
Instead of writing a list of app names that must be ignored, a regex can
be used to indicate either a list of names that must be ignored (just as
before, e.g. '^(?=.*(wine|yakuake))(.*)$') or a list of names that must
not be ignored (e.g. '^(?=.*(konsole|zathura))(.*)$'). In the second
case, every other program will be ignored.

Perhaps adding this as an alternative option instead of replacing the
current list of names may be a better/more flexible choice.